# perf: speed up property definitions query