--- a/latest_migrations.manifest
+++ b/latest_migrations.manifest
@@ -3,6 +3,6 @@ auth: 0012_alter_user_first_name_max_length
 axes: 0006_remove_accesslog_trusted
 contenttypes: 0002_remove_content_type_name
 ee: 0014_roles_memberships_and_resource_access
-posthog: 0296_team_allow_blank_fields
+posthog: 0297_property_definitions_index_query
 sessions: 0001_initial
 social_django: 0010_uid_db_index