Issue: https://github.com/PostHog/posthog/issues/13767,

Related PRs: https://github.com/PostHog/posthog/pull/13809, https://github.com/PostHog/posthog/pull/13816 https://github.com/PostHog/posthog/pull/13875 https://github.com/PostHog/posthog/pull/13916

This change adds the following index and starts using it

```
CREATE INDEX CONCURRENTLY "index_property_def_query" ON "posthog_propertydefinition" ("team_id", "type", (COALESCE("group_type_index",  -1)), "query_usage_30_day" DESC NULLS LAST, "name" ASC);
```

This speeds up property definitions query significantly.

New query plan:

```
+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| QUERY PLAN                                                                                                                                                                                                   |
|--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
| Limit  (cost=2618.85..2618.85 rows=1 width=131) (actual time=0.073..0.074 rows=0 loops=1)                                                                                                                    |
|   ->  Sort  (cost=2618.85..2618.85 rows=1 width=131) (actual time=0.073..0.074 rows=0 loops=1)                                                                                                               |
|         Sort Key: ((posthog_eventproperty.property IS NOT NULL)) DESC, dupe_propertydefinition.query_usage_30_day DESC NULLS LAST, dupe_propertydefinition.name                                              |
|         Sort Method: quicksort  Memory: 25kB                                                                                                                                                                 |
|         ->  Nested Loop Left Join  (cost=0.25..2618.84 rows=1 width=131) (actual time=0.067..0.068 rows=0 loops=1)                                                                                           |
|               Join Filter: (dupe_propertydefinition.id = ee_enterprisepropertydefinition.propertydefinition_ptr_id)                                                                                          |
|               ->  Nested Loop  (cost=0.25..2595.79 rows=1 width=113) (actual time=0.067..0.068 rows=0 loops=1)                                                                                               |
|                     ->  Unique  (cost=0.14..4.25 rows=21 width=14) (actual time=0.067..0.067 rows=0 loops=1)                                                                                                 |
|                           ->  Index Only Scan using posthog_event_property_unique_team_event_property on posthog_eventproperty  (cost=0.14..4.24 rows=21 width=14) (actual time=0.067..0.067 rows=0 loops=1) |
|                                 Index Cond: ((team_id = 1) AND (event = '$pageview'::text))                                                                                                                  |
|                                 Heap Fetches: 0                                                                                                                                                              |
|                     ->  Index Scan using test_idx on dupe_propertydefinition  (cost=0.11..123.40 rows=1 width=99) (never executed)                                                                           |
|                           Index Cond: ((team_id = 2) AND (type = 1) AND (COALESCE((group_type_index)::integer, '-1'::integer) = '-1'::integer) AND ((name)::text = (posthog_eventproperty.property)::text))  |
|                           Filter: (((name)::text <> '$foo'::text) AND ((name)::text ~~* '%$%'::text) AND ((name)::text !~~ '$feature/%'::text))                                                              |
|               ->  Seq Scan on ee_enterprisepropertydefinition  (cost=0.00..18.18 rows=1393 width=31) (never executed)                                                                                        |
| Planning Time: 0.668 ms                                                                                                                                                                                      |
| Execution Time: 0.130 ms                                                                                                                                                                                     |
+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
```

Old query plan:
```
+----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------->
| QUERY PLAN                                                                                                                                                                                                      >
|----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------->
| Limit  (cost=53860.29..53871.09 rows=100 width=129) (actual time=789.018..803.066 rows=0 loops=1)                                                                                                               >
|   ->  Gather Merge  (cost=53860.29..77214.39 rows=216348 width=129) (actual time=789.018..803.064 rows=0 loops=1)                                                                                               >
|         Workers Planned: 4                                                                                                                                                                                      >
|         Workers Launched: 4                                                                                                                                                                                     >
|         ->  Sort  (cost=52860.28..52887.32 rows=54087 width=129) (actual time=783.054..783.058 rows=0 loops=5)                                                                                                  >
|               Sort Key: (CASE WHEN (check_for_matching_event_property.id IS NULL) THEN false ELSE true END) DESC, posthog_propertydefinition.query_usage_30_day DESC NULLS LAST, posthog_propertydefinition.name>
|               Sort Method: quicksort  Memory: 25kB                                                                                                                                                              >
|               Worker 0:  Sort Method: quicksort  Memory: 25kB                                                                                                                                                   >
|               Worker 1:  Sort Method: quicksort  Memory: 25kB                                                                                                                                                   >
|               Worker 2:  Sort Method: quicksort  Memory: 25kB                                                                                                                                                   >
|               Worker 3:  Sort Method: quicksort  Memory: 25kB                                                                                                                                                   >
|               ->  Hash Left Join  (cost=66.16..52446.85 rows=54087 width=129) (actual time=782.967..782.970 rows=0 loops=5)                                                                                     >
|                     Hash Cond: (posthog_propertydefinition.id = ee_enterprisepropertydefinition.propertydefinition_ptr_id)                                                                                      >
|                     ->  Hash Left Join  (cost=43.10..52395.40 rows=54087 width=101) (actual time=782.966..782.969 rows=0 loops=5)                                                                               >
|                           Hash Cond: ((posthog_propertydefinition.name)::text = (check_for_matching_event_property.property)::text)                                                                             >
|                           Filter: CASE WHEN (check_for_matching_event_property.id IS NULL) THEN false ELSE true END                                                                                             >
|                           Rows Removed by Filter: 85188                                                                                                                                                         >
|                           ->  Parallel Seq Scan on posthog_propertydefinition  (cost=0.00..52271.16 rows=108174 width=97) (actual time=0.072..770.376 rows=85188 loops=5)                                       >
|                                 Filter: (((name)::text ~~* '%$%'::text) AND ((name)::text !~~ '$feature/%'::text) AND (team_id = 2) AND (type = 1))                                                             >
|                                 Rows Removed by Filter: 599127                                                                                                                                                  >
|                           ->  Hash  (cost=43.10..43.10 rows=1 width=18) (actual time=0.340..0.342 rows=0 loops=5)                                                                                               >
|                                 Buckets: 1024  Batches: 1  Memory Usage: 8kB                                                                                                                                    >
|                                 ->  Subquery Scan on check_for_matching_event_property  (cost=0.14..43.10 rows=1 width=18) (actual time=0.339..0.341 rows=0 loops=5)                                            >
|                                       ->  GroupAggregate  (cost=0.14..43.10 rows=1 width=54) (actual time=0.339..0.340 rows=0 loops=5)                                                                          >
|                                             Group Key: posthog_eventproperty.team_id, posthog_eventproperty.property                                                                                            >
|                                             Filter: (min((posthog_eventproperty.event)::text) = '$pageview'::text)                                                                                              >
|                                             ->  Index Scan using posthog_event_property_unique_team_event_property on posthog_eventproperty  (cost=0.14..42.98 rows=21 width=66) (actual time=0.339..0.339 rows=>
|                                                   Index Cond: ((team_id = 1) AND ((event)::text = '$pageview'::text))                                                                                           >
|                     ->  Hash  (cost=18.18..18.18 rows=1393 width=31) (never executed)                                                                                                                           >
|                           ->  Seq Scan on ee_enterprisepropertydefinition  (cost=0.00..18.18 rows=1393 width=31) (never executed)                                                                               >
| Planning Time: 3.149 ms                                                                                                                                                                                         >
| Execution Time: 803.246 ms                                                                                                                                                                                      >
+----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------->
```

Learning was that the old query would not use the trigram index in a majority of cases (e.g. when the filter was empty-ish) so the solution is to make everything else (filtering to relevant property definitions) as cheap as possible.